# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HRMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/attendance:
    get:
      operationId: getAttendanceReport
      tags: [reports]
      summary: Get attendance report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: departmentId
          schema:
            type: integer
      responses:
        "200":
          description: Attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceReport"
  /reports/leave:
    get:
      operationId: getLeaveReport
      tags: [reports]
      summary: Get leave report
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Leave report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveReport"

  # GEO-FENCE
...
components:
  schemas:
...
    AttendanceReport:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        summary:
          type: object
          properties:
            totalWorkingDays:
              type: integer
            averageAttendance:
              type: number
            totalPresent:
              type: integer
            totalAbsent:
              type: integer
            totalLate:
              type: integer
            totalOvertime:
              type: number
          required:
            - totalWorkingDays
            - averageAttendance
            - totalPresent
            - totalAbsent
            - totalLate
            - totalOvertime
        dailyBreakdown:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              present:
                type: integer
              absent:
                type: integer
              late:
                type: integer
            required:
              - date
              - present
              - absent
              - late
        employeeBreakdown:
          type: array
          items:
            type: object
            properties:
              employeeId:
                type: integer
              employeeName:
                type: string
              presentDays:
                type: integer
              absentDays:
                type: integer
              lateDays:
                type: integer
              attendancePercentage:
                type: number
            required:
              - employeeId
              - employeeName
              - presentDays
              - absentDays
              - lateDays
              - attendancePercentage
      required:
        - startDate
        - endDate
        - summary
        - dailyBreakdown
        - employeeBreakdown
    LeaveReport:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalRequests:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer
        byType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
              days:
                type: number
            required:
              - type
              - count
              - days
      required:
        - startDate
        - endDate
        - totalRequests
        - approved
        - rejected
        - pending
        - byType
...